Registration and Login is Done(BackEnd)
Added alot of files for proper MVC structure of the code. Please read the comments in the codes and also in "BasahaKo.txt" for explainations. Kulang nlng for Login is for the Admin. In the Database Added Priviledge to know if the certain user is an admin or regular user.